Ensure that a session cookie is set before each request.
This session cookie serves as both a unique session ID, and as a user ID
for each user. Since we're promoting anonymity, there's no need for a
user to have an indivdual account--we don't care who they are, exactly,
just if they've been here before or not.

To make this a little more sticky than a normal cookie, we'll use
the browser's localstorage (if available) to store it also,
and will use it as the canonical version per request.

While this is vulernable to cookie tampering, we're not terribly
concerned: we track accounts only for gamification purposes (totaling
points), so there's really no risk in impersonation if someone does
manage to pull it off.