Add basic session ID generation.
The gist of this idea is that every user should get a unique session ID,
which will serve as their user ID. Since we want to promote anonymity,
we're not going to do anything fancier than this.

We should use localstorage to try to persist this longer than cookies
will generally last.