Add a MainController and a main#index method. These will serve as the index page of our application, and the controller where we serve HTML.
We're using a single-age application approach here, which means that we
only serve preliminary HTML and JS from this single endpoint.

Also adds a unit test to ensure that session information is set as we
anticipate. @group, there are some TODOs here in the tests--a good
exercise would be to create this. You'll want to look into "selenium
webdriver."